Implemented swiping
I merged the swipe demo into the main page. I also merged the separate
pages into one that works with dragend, so I deleted the redundant
templates. You should be able to navigate the pages by sqiping, using
the navbar at the top, or clicking the buttons on the main page.