Quandl data gathering draft
Generates the list of country codes, grabs the corresponding data from
the web and prints the table of marriage age for each country as far
back as 2005. Todo: Trim the tables to get only the most recent data,
match the three-letter Quandl country codes to the country codes already
on the website.